Modify the GithubProposalKey type to `string`
The "Checking validity of types" step in Next.js's build seems to have
an issue when trying to use `${string}__${string}` to index an object
although tsc and VSCode have no issues with it.

Changing the GitHubProposalKey type to just be a `string` for now until
Next.js can handle the type.