Use new strategy for rendering readmes
Instead of getting the README html from the GitHub API, we will now
receive it the raw markdown as a string and then manually convert it
into HTML and highlight it. The issue with the old approach was that
relative links inside the readme html that was returned from GitHub
would be incorrect. By doing the HTML conversion ourselves, we can have
more control and set a base URL for the relative links.