Remove extraneous API calls when fetching proposal
Before, fetching a single proposal involved fetching all the proposals
for a stage as well as fetching the GitHub repository details for each
individual proposal. Now, we only fetch the GitHub repository details
for the specific proposal instead of every proposal in the same stage.

Because of this, we no longer need to have that hacky avoidRateLimit()
function because we are not exceeding GitHub's secondary rate limits.